requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
fastapi==0.96.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pydantic==1.10.8
pytest==7.3.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
